% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ac.R
\name{calc.ac}
\alias{calc.ac}
\title{Calculate Covariate Balance}
\usage{
calc.ac(w, X, weights)
}
\arguments{
\item{w}{A vector exposure values across all subjects.}

\item{X}{A matrix of covariate values. Subjects in rows and covariates in columns.}

\item{weights}{A vector of weights assigned to all subjects based on the trained GP.}
}
\value{
A vector of correlations between w and each column of X.
}
\description{
Calculate weighted correlation between a list of covariates and an exposure.
Weights are defined by the covariance function of the GP.
}
\examples{

}
