% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_nn_fast.R
\name{deriv.nn.fast}
\alias{deriv.nn.fast}
\title{Calculate Derivatives of CERF for nnGP}
\usage{
\method{deriv}{nn.fast}(
  w,
  w.obs,
  GPS.obs,
  y.obs,
  param,
  e_gps_pred,
  e_gps_std,
  n.neighbours,
  expand,
  block.size,
  kernel.fn = function(x) exp(-x^2),
  kernel.deriv.fn = function(x, mu, sigma) (x - mu)/sigma^2
)
}
\arguments{
\item{w}{A scalar of exposure level of interest.}

\item{w.obs}{A vector of observed exposure levels of all samples.}

\item{GPS.obs}{A vector of GPS for all samples at the observed levels of exposure.}

\item{y.obs}{A vector of observed outcome values.}

\item{param}{A vector of hyper-parameters in the GP model.}

\item{e_gps_pred}{A vecotor of estimated conditional means of the exposure given covariates
for all samples.}

\item{e_gps_std}{A scalar of estimated conditional standard deviations of the exposure given covariates.}

\item{expand}{Scaling factor to determine the total number of nearest neighbours. The total is \code{2*expand*n.neighbour}.}

\item{block.size}{Number of samples included in a computation block. Mainly used to
balance the speed and memory requirement. Larger \code{block.size} is faster, but requires more memory.}

\item{kernel.fn}{The covariance function.}

\item{kernel.deriv.fn}{The partial derivative of the covariance function.}

\item{n.neighbour}{Number of nearest neighbours on one side (see also \code{expand}).}
}
\value{
A scalar of estimated derivative of CERF at \code{w} in nnGP.
}
\description{
Calculate the posterior mean of the derivative of CERF at a given
exposure level with nnGP.
}
