% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cerf_gp.R
\name{estimate_cerf_gp}
\alias{estimate_cerf_gp}
\title{Estimate the Conditional Exposure Response Function using Gaussian Process}
\usage{
estimate_cerf_gp(
  data,
  w,
  GPS_m,
  params,
  nthread = 1,
  kernel_fn = function(x) exp(-x^2)
)
}
\arguments{
\item{data}{A data.table of observation data.
\itemize{
\item Column 1: Outcome (Y)
\item Column 2: Exposure or treatment (w)
\item Column 3~m: Confounders (C)
}}

\item{w}{A vector of exposure level to compute CERF.}

\item{GPS_m}{A data.table of GPS vectors.
\itemize{
\item Column 1: GPS
\item Column 2: Prediction of exposure for covariate of each data sample (e_gps_pred).
\item Column 3: Standard deviation of  e_gps (e_gps_std)
}}

\item{params}{A list of parameters that is required to run the process.
These parameters include:
\itemize{
\item alpha: A scaling factor for the GPS value.
\item beta: A scaling factor for the exposure value.
\item g_sigma: A scaling factor for kernel function (gamma/sigma).
\item tune_app: A tuning approach. Available approaches:
\itemize{
\item all: try all combinations of hyperparameters.
alpha, beta, and g_sigma can be a vector of parameters.
}
}}

\item{nthread}{An integer value that represents the number of threads to be
used by internal packages.}

\item{kernel_fn}{A kernel function. A default value is a Gaussian Kernel.}
}
\value{
A cerf_gp object that includes the following values:
\itemize{
\item TBD
}
}
\description{
Estimates the conditional exposure response function (cerf) using Gaussian
Process (gp). The function tune the best match (the lowest covariate balance)
for the provided set of hyperparameters.
}
\examples{

set.seed(129)
sim.data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)


# Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(sim.data[,-(1:2)]),
                   w.all = as.matrix(sim.data$treat))

# exposure values
w.all = seq(0,20,1)

data.table::setDT(sim.data)

cerf_gp_obj <- estimate_cerf_gp(sim.data,
                                w.all,
                                GPS_m,
                                params = list(alpha = c(0.1),
                                              beta=0.2,
                                              g_sigma = 1,
                                              tune_app = "all"),
                                nthread = 1)


}
