% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_noise_nn.R
\name{estimate_noise_nn}
\alias{estimate_noise_nn}
\title{Estimate the Standard Deviation (noise) of the Nugget Term in nnGP}
\usage{
estimate_noise_nn(hyperparam, w_obs, GPS_obs, y_obs, n_neighbor)
}
\arguments{
\item{hyperparam}{A vector of hyper-parameter values.}

\item{w_obs}{A vector of observed exposure levels.}

\item{GPS_obs}{A vector of estimated GPS evaluated at the observed exposure levels.}

\item{y_obs}{A vector of observed outcomes.}

\item{n_neighbor}{Number of nearest neighbors on one side.}
}
\value{
A scalar of estimated standard deviation of the nugget term in nnGP.
}
\description{
Estimate the standard deviations of the nugget term (noise) in nnGP by
calculating the standard deviations of the residuals.
}
\examples{

set.seed(425)
data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)

# Estimate GPS function
GPS_m <- train_GPS(cov_mt = data[,-(1:2)], w_all = data$treat)

# Hyperparameter
hyperparam <- c(0.1, 0.2, 1)
n_neighbor <- 10
expand <- 1
block_size <- 10000

# Exposure level
wi <- 1.2

# Estimate GPS for the exposure level
GPS_w = dnorm(wi,
              mean = GPS_m$e_gps_pred,
              sd = GPS_m$e_gps_std, log = TRUE)

# Order data for easy selection
coord_obs = cbind(data$treat, GPS_m$GPS)
y_use <- data$Y

obs_ord <- coord_obs[order(coord_obs[,1]),]
y_use_ord <- y_use[order(coord_obs[,1])]

noise <- estimate_noise_nn(hyperparam = hyperparam,
                           w_obs = data$treat,
                           GPS_obs = GPS_m$GPS,
                           y_obs = y_use_ord,
                           n_neighbor = n_neighbor)

}
