% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nn_fast.R
\name{get.nn.fast}
\alias{get.nn.fast}
\title{Title}
\usage{
get.nn.fast(
  params,
  w.new,
  GPS.new,
  obs.ord,
  y.obs.ord,
  n.neighbour = 10,
  expand = 5,
  block.size = 10000
)
}
\arguments{
\item{params}{param's description}

\item{w.new}{param's description}

\item{GPS.new}{param's description}

\item{obs.ord}{param's description}

\item{y.obs.ord}{param's description}

\item{n.neighbour}{param's description}

\item{expand}{param's description}

\item{block.size}{param's description}
}
\value{

}
\description{
Description
}
