% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_cp_diff.R
\name{GP.cp.diff}
\alias{GP.cp.diff}
\title{Change-point Detection in Full GP}
\usage{
GP.cp.diff(
  w,
  y.obs,
  w.obs,
  GPS.obs,
  param,
  e_gps_pred,
  e_gps_std,
  kernel.fn,
  kernel.deriv.fn
)
}
\arguments{
\item{w}{A numeric value of the exposure level at which the difference between two one-sided
partial derivatives is calculated.}

\item{y.obs}{A vector of observed outcomes.}

\item{w.obs}{A vector of observed exposure levels.}

\item{GPS.obs}{A vector of estimated GPS at the observed exposure levels.}

\item{param}{A vector of hyperparameters for the GP.}

\item{e_gps_pred}{A vector of estimated conditional means of the exposure given covariates
for all samples.}

\item{e_gps_std}{A scalar of estimated conditional standard deviations of the exposure given covariates.}

\item{kernel.fn}{The covariance function of GP.}

\item{kernel.deriv.fn}{The partial derivative of the covariance function.}
}
\value{
A numeric value of the posterior mean of the difference between two one-sided
derivatives.
}
\description{
Calculate the posterior mean of the difference between left- and right-derivatives
at an exposure level for the detection of change points.
}
