% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_weights_test.R
\name{GP.weights.test}
\alias{GP.weights.test}
\title{Title}
\usage{
GP.weights.test(
  w,
  w.obs,
  obs.use,
  param,
  inv.Sigma.obs,
  e_gps_pred,
  e_gps_std,
  kernel.fn = function(x) exp(-x^2)
)
}
\arguments{
\item{kernel.fn}{}
}
\value{

}
\description{

}
