% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_cp_calc.R
\name{nn.cp.calc}
\alias{nn.cp.calc}
\title{Change-point Detection in nnGP}
\usage{
nn.cp.calc(
  w,
  w.obs,
  GPS.obs,
  y.obs,
  param,
  e_gps_pred,
  e_gps_std,
  kernel.fn = function(x) exp(-x^2),
  kernel.deriv.fn = function(x, mu, sigma) (x - mu)/sigma^2,
  n.neighbour,
  expand,
  block.size
)
}
\arguments{
\item{w}{A numeric value of the exposure level at which the difference between two one-sided
partial derivatives is calculated.}

\item{w.obs}{A vector of observed exposure levels.}

\item{GPS.obs}{A vector of estimated GPS at the observed exposure levels.}

\item{y.obs}{A vector of observed outcomes.}

\item{param}{A vector of hyperparameters for the GP.}

\item{e_gps_pred}{A vector of estimated conditional means of the exposure given covariates
for all samples.}

\item{e_gps_std}{A scalar of estimated conditional standard deviations of the exposure given covariates.}

\item{kernel.fn}{The covariance function of GP.}

\item{kernel.deriv.fn}{The partial derivative of the covariance function.}

\item{n.neighbour}{Number of nearest neighbours on one side (see also \code{expand}).}

\item{expand}{Scaling factor to determine the total number of nearest neighbours. The total is \code{2*expand*n.neighbour}.}
}
\value{
A numeric value of the posterior mean of the difference between two one-sided
derivatives.
}
\description{
Calculate the posterior mean of the difference between left- and right-derivatives
at an exposure level for the detection of change points. nnGP approximation is used.
}
