% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_sigma_est.R
\name{nn.sigma.est}
\alias{nn.sigma.est}
\title{Title}
\usage{
nn.sigma.est(params, w.obs, GPS.obs, y.obs, n.neighbour, n.core = 20)
}
\arguments{
\item{params}{param's description}

\item{w.obs}{param's description}

\item{GPS.obs}{param's description}

\item{y.obs}{param's description}

\item{n.neighbour}{param's description}

\item{n.core}{param's description}
}
\value{

}
\description{
Description
}
