% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_est.R
\name{noise.est}
\alias{noise.est}
\title{Estimate the Standard Deviation of the Nugget Term in Full Gaussian Process}
\usage{
noise.est(param, data, GPS)
}
\arguments{
\item{param}{A vector of hyper-parameter values for the full GP.}

\item{data}{A data frame containing outcome, exposure and all covariates.}

\item{GPS}{A vector of estimated GPS at the observed exposure levels.}
}
\value{
A scalar of estimated standard deviation of the nugget term in full GP.
}
\description{
Estimates the standard deviations of the nugget term in full GP by calculating
the standard deviations of the residuals.
}
\examples{

}
