% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tru_R.R
\name{tru_R}
\alias{tru_R}
\title{Calculate True CERF}
\usage{
tru_R(w, sim.data)
}
\arguments{
\item{w}{A exposure level at which the CERF is derived.}

\item{sim.data}{A data frame from one simulation run.}
}
\value{
The CERF value evaluated at \code{w}.
}
\description{
Derives the true CERF at exposure levels of interest when the data
is distributed as described in the simulation study.
}
\examples{

sim.data <- generate_synthetic_data()
w <- 0.2
tr <- tru_R(w, sim.data)

}
