% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ac.R
\name{calc.ac}
\alias{calc.ac}
\title{Title}
\usage{
calc.ac(w, X, weights)
}
\arguments{
\item{w}{param description}

\item{X}{param description}

\item{weights}{param description}
}
\value{
Return val
}
\description{
Description
}
\examples{

}
