% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ac.R
\name{calc_ac}
\alias{calc_ac}
\title{Calculate covariate balance}
\usage{
calc_ac(w, X, weights)
}
\arguments{
\item{w}{A vector of exposure values across all subjects.}

\item{X}{A matrix of covariate values. Subjects in rows and covariates in columns.}

\item{weights}{A vector of weights assigned to all subjects based on the trained GP.}
}
\value{
A vector of correlations between w and each column of X.
}
\description{
Calculates the weighted correlation between a list of covariates and an
exposure. The covariance function of the GP defines weights.
}
\keyword{internal}
