% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{data_generate}
\alias{data_generate}
\title{Title}
\usage{
data_generate(sample_size = 1000, outcome_sd = 10, gps_spec = 1, cova_spec = 1)
}
\arguments{
\item{cova_spec}{}
}
\value{

}
\description{

}
