% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nn_fast.R
\name{get.nn.fast}
\alias{get.nn.fast}
\title{Title}
\usage{
get.nn.fast(
  params,
  w.new,
  GPS.new,
  obs.ord,
  y.obs.ord,
  n.neighbour = 10,
  expand = 5,
  block.size = 10000
)
}
\arguments{
\item{block.size}{}
}
\value{

}
\description{

}
