% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nn_sd.R
\name{get.nn.sd}
\alias{get.nn.sd}
\title{Title}
\usage{
get.nn.sd(params, w.new, GPS.new, obs.ord, n.neighbour = 10, expand = 1)
}
\arguments{
\item{params}{param's description}

\item{w.new}{param's description}

\item{GPS.new}{param's description}

\item{obs.ord}{param's description}

\item{n.neighbour}{param's description}

\item{expand}{param's description}
}
\value{

}
\description{
Description
}
