% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_deriv_weights_calc.R
\name{GP.deriv.weights.calc}
\alias{GP.deriv.weights.calc}
\title{Title}
\usage{
GP.deriv.weights.calc(
  w,
  w.obs,
  GPS.obs,
  param,
  e_gps_pred,
  e_gps_std,
  kernel.fn = function(x) exp(-x^2),
  kernel.deriv.fn = function(x, mu, sigma) (x - mu)/sigma^2
)
}
\arguments{
\item{kernel.deriv.fn}{}
}
\value{

}
\description{

}
