% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_weights_test.R
\name{GP.weights.test}
\alias{GP.weights.test}
\title{Title}
\usage{
GP.weights.test(
  w,
  w.obs,
  obs.use,
  param,
  inv.Sigma.obs,
  e_gps_pred,
  e_gps_std,
  kernel.fn = function(x) exp(-x^2)
)
}
\arguments{
\item{w}{param's description}

\item{w.obs}{param's description}

\item{obs.use}{param's description}

\item{param}{param's description}

\item{inv.Sigma.obs}{param's description}

\item{e_gps_pred}{param's description}

\item{e_gps_std}{param's description}

\item{kernel.fn}{param's description}
}
\value{

}
\description{
Description
}
