% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_balance.R
\name{nn.balance}
\alias{nn.balance}
\title{Title}
\usage{
nn.balance(
  w.obs,
  w.est,
  y.obs,
  train.GPS.ret,
  design.mt,
  all.params = expand.grid(seq(0.5, 4.5, 1), seq(0.5, 4.5, 1), seq(0.5, 4.5, 1)),
  n.cpu = 20,
  n.neighbour = 50,
  expand = 2,
  block.size = 2000
)
}
\arguments{
\item{block.size}{}
}
\value{

}
\description{

}
