% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_estimate.R
\name{nn.estimate}
\alias{nn.estimate}
\title{Title}
\usage{
nn.estimate(
  params,
  w.obs,
  w.est,
  y.obs,
  train.GPS.ret,
  n.cpu = 20,
  n.neighbour = 50,
  expand = 2,
  block.size = 2000
)
}
\arguments{
\item{params}{param's description}

\item{w.obs}{param's description}

\item{w.est}{param's description}

\item{y.obs}{param's description}

\item{train.GPS.ret}{param's description}

\item{n.cpu}{param's description}

\item{n.neighbour}{param's description}

\item{expand}{param's description}

\item{block.size}{param's description}
}
\value{

}
\description{
Description
}
