% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_estimate.R
\name{nn.estimate}
\alias{nn.estimate}
\title{Title}
\usage{
nn.estimate(
  params,
  w.obs,
  w.est,
  y.obs,
  train.GPS.ret,
  n.cpu = 20,
  n.neighbour = 50,
  expand = 2,
  block.size = 2000
)
}
\arguments{
\item{block.size}{}
}
\value{

}
\description{

}
