% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_est.R
\name{noise.est}
\alias{noise.est}
\title{Title}
\usage{
noise.est(param, sim.data, GPS)
}
\arguments{
\item{param}{param's description}

\item{sim.data}{param's description}

\item{GPS}{param's description}
}
\value{

}
\description{
Description
}
