% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_fn.R
\name{tuning.fn}
\alias{tuning.fn}
\title{Hyperparameter Tuning in Full GP}
\usage{
tuning.fn(
  param,
  sim.data,
  w.all,
  GPS,
  e_gps_pred,
  e_gps_std,
  kernel.fn = function(x) exp(-x^2)
)
}
\arguments{
\item{param}{A vector of values of hyperparameters.}

\item{sim.data}{A data frame containing all data including outcome, exposure
and covariates.}

\item{w.all}{A vector of exposure levels at which the CERF is estimated.}

\item{GPS}{A vector of estimated GPS evaluated at the observed exposure levels.}

\item{e_gps_pred}{A vector of estimated conditional means of the exposure given covariates
for all samples.}

\item{e_gps_std}{A scalar of estimated conditional standard deviations of the exposure given covariates.}

\item{kernel.fn}{The covariance function of GP.}
}
\value{
A list containing two elements: 1) a vector of absolute weighted correlation of each
covariate to the exposure, which is the metric for covariate balance and 2) the estimated
CERF at \code{w.all} based on the hyper-parameter values in \code{param}.
}
\description{
Calculate the induced covariate balance associated with one hyperparameter
configuration in full GP.
}
