% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_fn.R
\name{tuning.fn}
\alias{tuning.fn}
\title{Title}
\usage{
tuning.fn(
  param,
  sim.data,
  w.all,
  GPS,
  e_gps_pred,
  e_gps_std,
  kernel.fn = function(x) exp(-x^2)
)
}
\arguments{
\item{param}{param's description}

\item{sim.data}{param's description}

\item{w.all}{param's description}

\item{GPS}{param's description}

\item{e_gps_pred}{param's description}

\item{e_gps_std}{param's description}

\item{kernel.fn}{param's description}
}
\value{

}
\description{
Description
}
